Product_Pro/Product_app/serializers.py
from rest_framework import serializers
from Product_app.models import Products

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = '__all__'